feat: create @launchql/env package for consolidated environment management
- Create new @launchql/env package with same structure as @launchql/types
- Move walkUp utility from core/workspace/utils to env package
- Move environment parsing logic from types to env package
- Move config loading logic from core LaunchQLProject to env package
- Enhance getEnvOptions with config file support using proper merge hierarchy:
  defaults → config file → env vars → overrides
- Update all 25+ files to import getEnvOptions from @launchql/env instead of @launchql/types
- Maintain type safety by importing LaunchQLOptions from @launchql/types
- Package builds successfully with dual CommonJS/ESM output matching types package
