+++ packages/cli/src/commands/deploy.ts
@@ -1,6 +1,6 @@
 import { LaunchQLProject } from '@launchql/core';
 import { Logger } from '@launchql/logger';
-import { getEnvOptions } from '@launchql/types';
+import { getEnvOptions } from '@launchql/env';
 import { execSync } from 'child_process';
 import { CLIOptions, Inquirerer, Question } from 'inquirerer';
 import { ParsedArgs } from 'minimist';
