+++ packages/core/src/files/sql/writer.ts
@@ -1,4 +1,4 @@
-import { getEnvOptions } from '@launchql/types';
+import { getEnvOptions } from '@launchql/env';
 import fs from 'fs';
 import path from 'path';
