+++ packages/server/src/run.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import { getEnvOptions } from '@launchql/types';
+import { getEnvOptions } from '@launchql/env';
 
 import { LaunchQLServer as server } from './server';
