+++ sandbox/my-third/__tests__/deploy-fast.test.ts
@@ -1,6 +1,6 @@
 import { deployFast, LaunchQLProject } from '@launchql/core';
 import { resolve } from 'path';
-import { getEnvOptions } from '@launchql/types';
+import { getEnvOptions } from '@launchql/env';
 import { randomUUID } from 'crypto';
 import { execSync } from 'child_process';
 import { getPgPool } from 'pg-cache';
